Split TextInput and Select so they can be used with or without validation
